fix: move correction to common
